move test__mesh_to_MeshInfo.py, fix tests
