+++ esmf_regrid/tests/unit/experimental/io/test_round_tripping.py
@@ -9,7 +9,7 @@ from esmf_regrid.experimental.unstructured_scheme import (
     GridToMeshESMFRegridder,
     MeshToGridESMFRegridder,
 )
-from esmf_regrid.tests.unit.experimental.unstructured_scheme.test__mesh_to_MeshInfo import (
+from esmf_regrid.tests.unit.schemes.test__mesh_to_MeshInfo import (
     _gridlike_mesh_cube,
 )
 from esmf_regrid.tests.unit.schemes.test__cube_to_GridInfo import (
