+++ esmf_regrid/tests/unit/schemes/test_ESMFBilinear.py
@@ -5,7 +5,7 @@ from numpy import ma
 import pytest
 
 from esmf_regrid.schemes import ESMFBilinear
-from esmf_regrid.tests.unit.experimental.unstructured_scheme.test__mesh_to_MeshInfo import (
+from esmf_regrid.tests.unit.schemes.test__mesh_to_MeshInfo import (
     _gridlike_mesh_cube,
 )
 from esmf_regrid.tests.unit.schemes.test__cube_to_GridInfo import _grid_cube
